requirements.txt
fastapi
uvicorn
python-multipart
torch         # or torch if you choose PyTorch
opencv-python
numpy
pillow
scikit-image        # optional for image quality metrics
